Find orthologs between species

Convert entrez gene ID from one species to another using the DIOPT service. This function is a wrapper of the DIOPT API.

Usage

find_orthologs(genes, from = 9606, to = 7227, filter = "none", version = 8)

Arguments

genes

A chr vector of genes. See @details

from

Source species. See @details

to

Target species. See @details

filter

Filter of the results. See @details.

version

What version should be used. Currently v9 is still in beta.

Value

A tibble with requested information.

Details

The genes argument can either be a enterzID, which will be used as is. Or it can be a gene symbol, which will be converted to entrezID using the org.Hs.eg.db::org.Hs.eg.db database.

The from and to arguments are the NCBI Taxonomy ID. DIOPT currently supports the following IDs:

"all": All Species (Avaiable for Target only)

"4896": Schizosaccharomyces pombe (Fission yeast)

"4932": Saccharomyces cerevisiae (Yeast)

"6239": Caenorhabditis elegans (Worm)

"7227": Drosophila melanogaster (Fly)

"7955": Danio rerio (Zebrafish)

"8364": Xenopus tropicalis (Western clawed frog)

"9606": Homo sapiens (Human)

"10090": Mus musculus (Mouse)

"10116": Rattus norvegicus (Rat)

"3702": Arabidopsis thaliana (Thale cress)

If from/to is no the same length as the input genes or 1, will warning and recycle to match the length of gene.

The filter argument filter based on score. Options are: none/best_match/exclude_score_less_1/exclude_score_less_2. See API Documentation

The function is not vectorized, because DIOPT API is not vectorized.